+++ src/Vetuviem.SourceGenerator/AbstractBaseGenerator.cs
@@ -17,8 +17,10 @@ namespace Vetuviem.SourceGenerator
     /// Base logic for a source generator.
     /// </summary>
     /// <typeparam name="TGeneratorProcessor"></typeparam>
-    public abstract class AbstractBaseGenerator<TGeneratorProcessor> : ISourceGenerator
-        where TGeneratorProcessor : AbstractGeneratorProcessor, new()
+    /// <typeparam name="TClassGenerator"></typeparam>
+    public abstract class AbstractBaseGenerator<TGeneratorProcessor, TClassGenerator> : ISourceGenerator
+        where TGeneratorProcessor : AbstractGeneratorProcessor<TClassGenerator>, new()
+        where TClassGenerator : IClassGenerator, new()
     {
         /// <inheritdoc />
         public void Initialize(GeneratorInitializationContext context)
