+++ src/Vetuviem.SourceGenerator/AbstractViewBindingModelGenerator.cs
@@ -1,8 +1,9 @@
-﻿using Vetuviem.SourceGenerator.GeneratorProcessors;
+﻿using Vetuviem.SourceGenerator.Features.ViewBindingModels;
+using Vetuviem.SourceGenerator.GeneratorProcessors;
 
 namespace Vetuviem.SourceGenerator
 {
-    public abstract class AbstractViewBindingModelGenerator : AbstractBaseGenerator<ViewBindingModelGeneratorProcessor>
+    public abstract class AbstractViewBindingModelGenerator : AbstractBaseGenerator<ViewBindingModelGeneratorProcessor, ViewBindingModelClassGenerator>
     {
         protected override string GetNamespace() => $"ReactiveUI.{GetPlatformName()}.ViewToViewModelBindings";
     }
