+++ src/Vetuviem.SourceGenerator/Features/Core/AbstractClassGenerator.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+
+namespace Vetuviem.SourceGenerator.Features.Core
+{
+    public interface IClassGenerator
+    {
+        ClassDeclarationSyntax GenerateClass(
+            INamedTypeSymbol namedTypeSymbol,
+            string baseUiElement,
+            string desiredCommandInterface,
+            string platformName);
+    }
+}
